v0.23 by Tanoshizo
Version  v0.23 by Tanoshizo:

Add @Tanoshiizo#7629 Version:

- Added drone formation switch
-Added ship hability support
-Bot will no longer wait untill ship hp is lower than 98% to attack npcs in palla
-Bot will no longer switch configuration when collecting boxes, only for palladium
-Bot will wait until attack config shield is >= 90% before changing while collecting palla
-Removed "Delta option" and "Palladium logic" options (all the options that needed those will still work)

How to use drone formation switch and ship hability:
Both settings are in the extension options
you just need to toggle the options
and set the slot position for each item.

BOTH OPTIONS ONLY WORK IF YOUR SLOT IS SET TO NUMBERS 0 TO 9
and remember to check you game settings if the keys are right

How the ship hability feature works:
Venom - It will be use only on npcs with hp above 60000
Cyborg - Same as venom but 100k hp
Diminisher - Activates on npcs with shield above 60000
Spectrum - Activates when running from enemy
Sentinel - Activates while fighting npcs
Solace - Activates when hp is 70%

All those options are going to be made user-defined in the future.
This is only a beta and your feedback is more than welcome.

Don't forget to report bugs in #bug-reports

ps: there's a little surprise in one of the windows

- Added translations of new things.
- Fixed the failure of the new window
- The palladium logic has been set again. In the next version it will change.